Updated file to newer ROOT format
The file contains one histogram used during reconstruction. Unfortunately, the file was
written so long ago that ROOT's schema for the classes has changed. This is a problem for the
threaded framework since ROOT 5 updates schema in a thread-unsafe manner when it opens
a file and sees a new-to-it schema. By creating a new file containing the same information
we can successfully preload the new class schemas (since they match the C++ code) at job
initialization time.
ROOT 6 is supposed to fix this problem.