Do not force creation of StreamerInfos at ROOT file opening
Forcing the creating of StreamerInfos at file open is no longer needed
and it causes problems when we have done a 'drop on input' for a file
and the items dropped can not be schema evolved. In that case the original
code would cause an exception and it would kill the job. This change
allows us to properly handle 'drop on input'.