Make ProcForeach used by MVAComputer thread safe
ProcForeach was using mutable member data to keep track of the
state of the looping it was doing. Given that multiple threads
could be updating those values concurrently, this lead to crashes.
The code has now been changed to store the looping state externally
via VarProcessor::LoopCtx on the stack inside MVAComputer::evalInternal.
This allows each thread to have its own copy.