Add HitTripletEDProducerT and IntermediateHitTriplets
Produces hit triplets directly to edm::Event, either as a form of
vector<SeedingHitSet> for SeedCreator, or as IntermediateHitTriplets
for quadruplet generation.

Also have to bookkeep the 3rd layers where the triplets come from,
because the triplets (from a given layer pair) need to be ordered by
their 3rd layer for quadruplet generation.